fix: Eliminate all compiler warnings
- Fix missing network_config initializers in SimulatedCluster::Config
  - Added .network_config = {} to test files and benchmarks
  - Ensures all struct fields are explicitly initialized

- Fix unused variable warnings
  - Marked unused variables with (void) cast or [[maybe_unused]]
  - Preserved variables for future use or debugging

- Fix prometheus_exporter write() return value warnings
  - Changed from (void)write() to [[maybe_unused]] auto result = write()
  - Satisfies warn_unused_result attribute requirement

- Fix sign comparison warning in phase_5_chaos_engineering
  - Cast atomic<int> to size_t for comparison with getDroppedMessages()
  - Prevents signed/unsigned comparison warning

All 328 tests pass (100%)
Build completes with 0 warnings

+++ tests/e2e/phase_5_chaos_engineering.cpp
@@ -832,7 +832,7 @@ TEST_F(Phase5MessageLossTest, MessageLossWithManualRetries) {
       total_attempts++;
 
       // Simulate sending (network may drop it)
-      bool dropped = (network.getDroppedMessages() < total_attempts.load());
+      bool dropped = (network.getDroppedMessages() < static_cast<size_t>(total_attempts.load()));
 
       if (!dropped) {
         // Message delivered
